refactor: split api routes into seperate files
